@@ -7,15 +7,15 @@
     "test": "jest"
   },
   "dependencies": {
-    "react": "16.3.1",
-    "react-native": "0.55.4",
-    "react-native-code-push": "^5.3.3"
+    "react": "18.1.0",
+    "react-native": "0.70.3",
+    "react-native-code-push": "^7.0.5"
   },
   "devDependencies": {
-    "babel-jest": "23.6.0",
-    "babel-preset-react-native": "4.0.0",
-    "jest": "23.6.0",
-    "react-test-renderer": "16.3.1"
+    "babel-jest": "29.1.2",
+    "babel-preset-react-native": "4.0.1",
+    "jest": "29.1.2",
+    "react-test-renderer": "18.1.0"
   },
   "jest": {
     "preset": "react-native"
